% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_mer.R
\name{bootstrap_mer}
\alias{bootstrap_mer}
\title{Run Various Bootstrap for Mixed Models.}
\usage{
bootstrap_mer(
  x,
  FUN,
  nsim = 1,
  seed = NULL,
  type = c("parametric", "residual", "residual_cgr", "residual_trans", "reb", "case"),
  corrected_trans = FALSE,
  lv1_resample = FALSE,
  reb_scale = FALSE,
  .progress = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{A fitted \code{merMod} object from \code{\link[lme4]{lmer}}.}

\item{FUN}{A function taking a fitted \code{merMod} object as input and
returning the statistic of interest, which must be a (possibly named)
numeric vector.}

\item{nsim}{A positive integer telling the number of simulations, positive
integer; the bootstrap \eqn{R}.}

\item{seed}{Optional argument to \code{\link[base]{set.seed}}.}

\item{type}{A character string indicating the type of multilevel bootstrap.
Currently, possible values are \code{"parametric"}, \code{"residual"},
\code{"residual_cgr"}, \code{"residual_trans"}, \code{"reb"},
or \code{"case"}.}

\item{corrected_trans}{Logical indicating whether to use the correct
variance-covariance matrix of the residuals. If \code{FALSE}, use the
variance of \eqn{y}; if \code{TRUE}, use the variance of \eqn{y - X \hat
\beta}. Only used for \code{type = "residual_trans"}.}

\item{lv1_resample}{Logical indicating whether to sample with replacement
the level-1 units for each level-2 cluster. Only used for
\code{type = "case"}. Default is \code{FALSE}.}

\item{reb_scale}{Logical indicating whether to scale the residuals for the
random effect block bootstrap}

\item{.progress}{Logical indicating whether to display progress bar (using
\code{\link[utils]{txtProgressBar}}).}

\item{verbose}{Logical indicating if progress should print output.}

\item{...}{argument passed to .resid_resample.}
}
\value{
An object of S3 class \code{"boot"}, compatible with \pkg{boot}
  package's \code{\link[boot]{boot}()}. It contains the following components:

  \item{t0}{The original statistic from \code{FUN(x)}.}
  \item{t}{A matrix with \code{nsim} rows containing the bootstrap
  distribution of the statistic.}
  \item{R}{The value of \code{nsim} passed to the function.}
  \item{data}{The data used in the original analysis.}
  \item{seed}{The value of \code{.Random.seed} when \code{bootstrap_mer}
  started to work.}
  \item{statistic}{The function \code{FUN} passed to \code{bootstrap_mer}.}

See the documentation in for \code{link[boot]{boot}()} for the other
  components.
}
\description{
Run multilevel parametric, residual, and case bootstrap with different
  options
}
\details{
\code{bootstrap_mer} performs different bootstrapping methods to fitted
model objects using the \pkg{lme4} package. Currently, only models fitted
using \code{\link[lme4]{lmer}} is supported.
}
\examples{
library(lme4)
fm01ML <- lmer(Yield ~ (1 | Batch), Dyestuff, REML = FALSE)
mySumm <- function(x) {
  c(getME(x, "beta"), sigma(x))
}
# Covariance preserving residual bootstrap
boo01 <- bootstrap_mer(fm01ML, mySumm, type = "residual", nsim = 100)
# Plot bootstrap distribution of fixed effect
library(boot)
plot(boo01, index = 1)
# Get confidence interval
boot.ci(boo01, index = 2, type = c("norm", "basic", "perc"))
# BCa using influence values computed from `empinf_mer`
boot.ci(boo01, index = 2, type = "bca", L = empinf_mer(fm01ML, mySumm, 2))
}
\references{
Carpenter, J. R., Goldstein, H., & Rasbash, J. (2003). A novel
  bootstrap procedure for assessing the relationship between class size and
  achievement. Journal of the Royal Statistical Society. Series C (Applied
  Statistics), 52, 431–443. https://doi.org/10.1111/1467-9876.00415

Chambers, R., & Chandra, H. (2013). A random effect block
  bootstrap for clustered data. Journal of Computational and Graphical
  Statistics, 22(2), 452–470. https://doi.org/10.1080/10618600.2012.681216

Davison, A. C. and Hinkley, D. V. (1997). Bootstrap methods and
  their application. Cambridge, UK: Cambridge University Press.

Morris, J. S. (2002). The BLUPs are not "best" when it comes to
  bootstrapping. Statistics & Probability Letters, 56(4), 425–430.
  https://doi.org/10.1016/S0167-7152(02)00041-X

Van der Leeden, R., Meijer, E., & Busing, F. M. T. A. (2008).
  Resampling multilevel models. In J. de Leeuw & E. Meijer (Eds.), Handbook
  of multilevel Analysis (pp. 401–433). New York, NY: Springer.
}
\seealso{
\itemize{
  \item \code{\link[boot]{boot}} for single-level bootstrapping,
  \item \code{\link[lme4]{bootMer}} for parametric and semi-parametric
  bootstrap implemented in lme4, and
  \item \code{\link[boot]{boot.ci}} for getting bootstrap confidence
  intervals and \code{\link[boot]{plot.boot}} for plotting the bootstrap
  distribution.}
}
