% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_residuals.R
\name{solve_eigen_sqrt}
\alias{solve_eigen_sqrt}
\title{Get the square root of a matrix using eigenvalue decomposition, and solve
for a linear system}
\usage{
solve_eigen_sqrt(M, b)
}
\arguments{
\item{M}{a symmetric positive definite/semi-definite matrix}

\item{b}{a numeric vector or matrix}
}
\description{
Solve for \eqn{x} in the linear system
\eqn{Ax = b}, where \eqn{A} is a
symmetric matrix square root of \eqn{M} with eigenvalue
decomposition such that \eqn{AA = M}.
}
