% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov_vc.R
\name{vcov_theta}
\alias{vcov_theta}
\title{Asymptotic Covariance Matrix for Cholesky Factor of Random Effects}
\usage{
vcov_theta(x)
}
\arguments{
\item{x}{A fitted merMod object from \code{\link[lme4]{lmer}}.}
}
\description{
Asymptotic Covariance Matrix for Cholesky Factor of Random Effects
}
\seealso{
\code{\link{vcov_vc}}
}
