% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootmlm.R
\docType{package}
\name{bootmlm}
\alias{bootmlm}
\alias{bootmlm-package}
\title{bootmlm: A package for bootstrap resampling with multilevel data.}
\description{
The bootmlm package provides the function bootstrapMer, which performs
various parametric, residual, and case bootstrap resampling for
fitted model objects with the lme4 package.
}
\section{Parametric bootstrap}{

}

\section{Residual bootstrap}{

}

\section{Case bootstrap}{

}

