% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.boot}
\alias{confint.boot}
\title{Bootstrap confidence intervals for Two-Level Mixed Models}
\usage{
\method{confint}{boot}(
  object,
  parm,
  level = 0.95,
  type = c("norm", "basic", "perc", "bca"),
  L = NULL,
  ...
)
}
\arguments{
\item{object}{an object returned by \code{\link{bootstrap_mer}}.}

\item{parm}{a specification of which parameters are to be given confidence
intervals as a vector of numbers. If missing, all parameters
are considered.}

\item{level}{the confidence level required.}

\item{type}{character indicating the type of intervals required, as described
in \code{\link[boot]{boot.ci}}. Currently \code{"stud"} is not
supported.}

\item{L}{empirical influence values required for \code{type = "bca"} as
described in \code{\link[boot]{boot.ci}}.}

\item{...}{additional argument(s) passed to \code{\link[boot]{boot.ci}}.}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence
        limits for each parameter.
}
\description{
This is a wrapper for getting CIs for multiple parameters after running
\code{\link{bootstrap_mer}}, to be consistent with a similar method for the
\code{\link[lme4]{bootMber}} class.
}
