% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov_vc.R
\name{devfun_mer}
\alias{devfun_mer}
\alias{devfun_mer2}
\title{Deviance Function for Multilevel Models}
\usage{
devfun_mer(x)

devfun_mer2(x)
}
\arguments{
\item{x}{A fitted merMod object from \code{\link[lme4]{lmer}}.}
}
\value{
A deviance function with one argument \code{th_sig} that takes input
  of a numeric vector corresponding to the some estimated values of
  \eqn{\theta} and \eqn{\sigma}. For \code{devfun_mer2}, a function with
  one argument \code{theta} that profiles out \eqn{\sigma} and only takes
  input of elements for \eqn{\theta}.
}
\description{
Creates a deviance function for a fitted model object, using \eqn{\theta}
and \eqn{sigma} as the parameters.
}
\details{
The built-in function(s) in \pkg{lme4} for generating deviance function are
not exported and rely on C++ code. This function is mainly used to obtain
Hessian and asymptotic covariance matrix of the random effects.
}
\examples{
library(lme4)
fm01ML <- lmer(Yield ~ (1 | Batch), Dyestuff, REML = FALSE)
dd <- devfun_mer(fm01ML)
# Asymptotic variance-covariance matrix of (theta, sigma):
2 * solve(numDeriv::hessian(dd, c(fm01ML@theta, sigma(fm01ML))))
}
\references{
Bates, D., M\"{a}chler, M., Bolker, B. M., & Walker, S. C.
  Fitting linear mixed-effects models using lme4. Retrieved from
  \url{https://cran.r-project.org/web/packages/lme4/vignettes/lmer.pdf}

Implementation in \pkg{lme4pureR}:
  \url{https://github.com/lme4/lme4pureR/blob/master/R/pls.R}
}
